Updates cli structure
