+++ tests/cli/demultiplex/test_demultiplex_flowcell.py
@@ -5,7 +5,7 @@ from typing import Dict
 
 from cg.apps.demultiplex.demultiplex_api import DemultiplexingAPI
 from cg.apps.demultiplex.flowcell import Flowcell
-from cg.cli.demultiplex.base import demultiplex_flowcell
+from cg.cli.demultiplex.demux import demultiplex_flowcell
 from click import testing
